+++ src/main/java/com/app/repository/ChatRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.repository.query.Param;
 import com.app.modal.Chat;
 import com.app.modal.User;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
+public interface ChatRepository extends JpaRepository<Chat, Long> {
 
 	@Query("select c from Chat c join c.users u where u.id=:userId")
-	public List<Chat> findChatByUserId(Integer userId);
+	public List<Chat> findChatByUserId(Long userId);
 	
 	@Query("select c from Chat c Where c.is_group=false And :user Member of c.users And :reqUser Member of c.users")
 	public Chat findSingleChatByUsersId(@Param("user")User user, @Param("reqUser")User reqUser);
